opengl mode fixed and made into u32 in multiple nodes' indexed verts visualizations
camera fitting fixed in visualizer

deletion of visible keys with delete key implemented in visualizer

inf rays (rays that didn't hit) taken out of Trace node visualization

+++ Fissure/Camera/Camera.cpp
@@ -100,7 +100,7 @@ tbl raysToIdxVerts(LavaParams const* lp, tbl const& rays)
   iv("colors green") = &cg;
   iv("colors alpha") = &ca;
   iv("indices")      = &ind;
-  iv("mode")         = 1;            // 0 should be points, 1 should be lines
+  iv("mode")         = (u32)1;            // 0 should be points, 1 should be lines
   iv("type")         = tbl::StrToInt("IdxVerts");
   iv.flatten();
